requirements.txt
python-dateutil
pydantic
discord
yt_dlp
yandex_music
discord.py[voice]
PyNaCl